make reconnect a method of NetworkState
