+++ chain/network/src/peer_manager/tests/tier2.rs
@@ -1,10 +1,9 @@
 use crate::broadcast;
 use crate::network_protocol::testonly as data;
 use crate::peer_manager::connection_store::STORED_CONNECTIONS_MIN_DURATION;
+use crate::peer_manager::network_state::RECONNECT_ATTEMPT_INTERVAL;
 use crate::peer_manager::peer_manager_actor::Event as PME;
-use crate::peer_manager::peer_manager_actor::{
-    POLL_CONNECTION_STORE_INTERVAL, RECONNECT_ATTEMPT_INTERVAL,
-};
+use crate::peer_manager::peer_manager_actor::POLL_CONNECTION_STORE_INTERVAL;
 use crate::peer_manager::testonly::start as start_pm;
 use crate::peer_manager::testonly::ActorHandler;
 use crate::peer_manager::testonly::Event;
